requirements.txt
tenseal>=0.3.14
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
matplotlib>=3.7.0
tqdm>=4.66.0
pyyaml>=6.0
